+++ androidx/src/main/java/me/panpf/androidx/view/Displayx.java
@@ -22,7 +22,10 @@ import android.content.res.Configuration;
 import android.graphics.Point;
 import android.support.annotation.NonNull;
 import android.util.DisplayMetrics;
+import android.view.Surface;
 import android.view.WindowManager;
+
+import org.jetbrains.annotations.NotNull;
 
 public class Displayx {
 
@@ -65,6 +68,24 @@ public class Displayx {
 
     public static int getDensityDpi(@NonNull Context context) {
         return context.getResources().getDisplayMetrics().densityDpi;
+    }
+
+    /**
+     * Get the current window direction
+     */
+    public static int getDisplayRotation(@NotNull Activity activity) {
+        switch (activity.getWindowManager().getDefaultDisplay().getRotation()) {
+            case Surface.ROTATION_0:
+                return 0;
+            case Surface.ROTATION_90:
+                return 90;
+            case Surface.ROTATION_180:
+                return 180;
+            case Surface.ROTATION_270:
+                return 270;
+            default:
+                return 0;
+        }
     }